+++ src/main/java/miyucomics/hexical/casting/patterns/OpMagicMissile.kt
@@ -1,8 +1,11 @@
 package miyucomics.hexical.casting.patterns
 
 import at.petrak.hexcasting.api.misc.MediaConstants
-import at.petrak.hexcasting.api.spell.*
+import at.petrak.hexcasting.api.spell.ParticleSpray
+import at.petrak.hexcasting.api.spell.RenderedSpell
+import at.petrak.hexcasting.api.spell.SpellAction
 import at.petrak.hexcasting.api.spell.casting.CastingContext
+import at.petrak.hexcasting.api.spell.getVec3
 import at.petrak.hexcasting.api.spell.iota.Iota
 import miyucomics.hexical.entities.MagicMissileEntity
 import miyucomics.hexical.registry.HexicalEntities