+++ src/main/java/miyucomics/hexical/casting/patterns/getters/OpGetItemStackData.kt
@@ -3,14 +3,16 @@ package miyucomics.hexical.casting.patterns.getters
 import at.petrak.hexcasting.api.spell.ConstMediaAction
 import at.petrak.hexcasting.api.spell.casting.CastingContext
 import at.petrak.hexcasting.api.spell.getItemEntity
-import at.petrak.hexcasting.api.spell.iota.*
+import at.petrak.hexcasting.api.spell.iota.BooleanIota
+import at.petrak.hexcasting.api.spell.iota.DoubleIota
+import at.petrak.hexcasting.api.spell.iota.Iota
+import at.petrak.hexcasting.api.spell.iota.ListIota
 import miyucomics.hexical.casting.iota.IdentifierIota
 import net.minecraft.enchantment.EnchantmentHelper
 import net.minecraft.item.EnchantedBookItem
 import net.minecraft.item.ItemStack
 import net.minecraft.item.Items
 import net.minecraft.util.registry.Registry
-import java.lang.IllegalStateException
 
 class OpGetItemStackData(private val mode: Int) : ConstMediaAction {
 	override val argc = 1