+++ src/main/java/miyucomics/hexical/casting/patterns/getters/OpGetLivingEntityData.kt
@@ -3,10 +3,12 @@ package miyucomics.hexical.casting.patterns.getters
 import at.petrak.hexcasting.api.spell.ConstMediaAction
 import at.petrak.hexcasting.api.spell.casting.CastingContext
 import at.petrak.hexcasting.api.spell.getLivingEntityButNotArmorStand
-import at.petrak.hexcasting.api.spell.iota.*
+import at.petrak.hexcasting.api.spell.iota.BooleanIota
+import at.petrak.hexcasting.api.spell.iota.DoubleIota
+import at.petrak.hexcasting.api.spell.iota.Iota
+import at.petrak.hexcasting.api.spell.iota.ListIota
 import miyucomics.hexical.casting.iota.IdentifierIota
 import net.minecraft.util.registry.Registry
-import java.lang.IllegalStateException
 
 class OpGetLivingEntityData(private val mode: Int) : ConstMediaAction {
 	override val argc = 1