+++ src/main/java/miyucomics/hexical/casting/patterns/getters/OpGetPlayerData.kt
@@ -8,7 +8,6 @@ import at.petrak.hexcasting.api.spell.iota.Iota
 import at.petrak.hexcasting.api.spell.iota.NullIota
 import miyucomics.hexical.casting.iota.IdentifierIota
 import net.minecraft.util.registry.Registry
-import java.lang.IllegalStateException
 
 class OpGetPlayerData(private val mode: Int) : ConstMediaAction {
 	override val argc = 1