+++ src/main/java/miyucomics/hexical/casting/patterns/specks/OpIotaSpeck.kt
@@ -1,7 +1,8 @@
 package miyucomics.hexical.casting.patterns.specks
 
-import at.petrak.hexcasting.api.spell.*
+import at.petrak.hexcasting.api.spell.ConstMediaAction
 import at.petrak.hexcasting.api.spell.casting.CastingContext
+import at.petrak.hexcasting.api.spell.getEntity
 import at.petrak.hexcasting.api.spell.iota.Iota
 import at.petrak.hexcasting.api.spell.mishaps.MishapBadEntity
 import miyucomics.hexical.entities.SpeckEntity