+++ src/main/java/miyucomics/hexical/state/EvokeState.kt
@@ -1,6 +1,6 @@
 package miyucomics.hexical.state
 
-import java.util.UUID
+import java.util.*
 
 object EvokeState {
 	val active: HashMap<UUID, Boolean> = HashMap()